atc: structure: split out pool tests
concourse/concourse#5421

It was a bit distracting to exercise the thread-safety and Acquire/Release
functions in a file named after the policy interface.
